+++ include/pfnet/heur_PVPQ.h
@@ -12,15 +12,15 @@
 #define __HEUR_PVPQ_HEADER__
 
 #include <math.h>
-#include "net.h"
 #include "heur.h"
 
 #define HEUR_PVPQ_DEBUG FALSE
 
 // Function prototypes
-void HEUR_PVPQ_init(Heur* h, Net* net);
-void HEUR_PVPQ_clear(Heur* h, Net* net);
-void HEUR_PVPQ_apply_step(Heur* h, Constr* clist, Net* net, Branch* br, int t, Vec* var_values);
+Heur* HEUR_PVPQ_new(Net* net);
+void HEUR_PVPQ_init(Heur* h);
+void HEUR_PVPQ_clear(Heur* h);
+void HEUR_PVPQ_apply_step(Heur* h, Constr* clist, Branch* br, int t, Vec* var_values);
 void HEUR_PVPQ_free(Heur* h);
 
 #endif