+++ include/pfnet/pfnet.h
@@ -54,5 +54,9 @@
 #include "constr_AC_LIN_FLOW_LIM.h"
 #include "constr_BAT_DYN.h"
 #include "constr_LOAD_PF.h"
+#include "constr_CFUNC.h"
+
+// Heuristics
+#include "heur_PVPQ.h"
 
 #endif